refactor: views.py を 4 ファイル + 共通モジュールに分割（800 行規約準拠）
リポジトリ規約「800 行上限 / many small files」に従い、852 行に肥大化
した features/talk/views.py をパッケージ化して責務ごとに分割。

新構成:
- views/__init__.py     (28行) ビュークラスを re-export、urls.py からの
  既存 import パスは無変更
- views/_common.py      (148行) handle_synthesis_error / get_owned_session
  / create_assistant_message / attach_audio_to_message / SESSION_MAX_MESSAGES
- views/synthesize.py   (256行) TalkSynthesizeView / TodayInfoView /
  ConfigsListView（既存 talk/synthesize 系）
- views/sessions.py     (120行) ChatSessionListCreateView /
  ChatSessionDetailView / SessionPagination
- views/messages.py     (269行) ChatSessionMessageView /
  ChatSessionMessageEditView
- views/audio.py        (87行)  ChatSessionMessageAudioView /
  ChatSessionAudioBulkDeleteView

テスト追従:
- @patch("features.talk.views.TalkService") を具体モジュール
  (features.talk.views.messages.TalkService または .synthesize.TalkService)
  に変更（25 ヶ所）
- @patch("features.talk.views.HolidayClient" / .datetime) も
  features.talk.views.synthesize.* に変更

挙動・公開 API は変わらず、全 talk テスト 162 件 pass、ruff/format OK。
